refactor: extract chart from results
