+++ frontend/src/features/power_unit/PowerUnitResults.tsx
@@ -1,7 +1,4 @@
-import * as THREE from "three";
-import { useEffect, useMemo, useRef, useState } from "react";
-import { Canvas, useFrame, ThreeElements } from "@react-three/fiber";
-import { Stats, OrbitControls, useHelper, Html } from "@react-three/drei";
+import { useEffect, useMemo, useState } from "react";
 import { cp } from "../../data/cp";
 import {
   generate_verts,
@@ -17,126 +14,8 @@ import { usePower } from "./hooks/usePower";
 import { usePropellerStore } from "./stores/usePropeller";
 import { useEngineStore } from "./hooks/useEngine";
 import { barycentricAngle } from "../../utils/interpolation/binarySearch";
-import { verts } from "../../data/verts";
 import { eff } from "../../data/eff";
-
-interface SurfaceProps {
-  cpMarkers: Float32Array;
-}
-
-const Surface = ({ cpMarkers }: SurfaceProps) => {
-  const mesh = useRef<THREE.Mesh>(null!);
-  const points = useRef<THREE.Points>(null!);
-  const positionsRef = useRef<THREE.BufferAttribute>(null);
-  const surfacePositionsRef = useRef<THREE.BufferAttribute>(null);
-
-  const blades = usePropellerStore((state) => state.blades);
-
-  const chartType = "cp";
-
-  const center = useMemo(() => {
-    if (mesh.current) {
-      mesh.current.geometry.computeBoundingBox();
-      const center = new THREE.Vector3();
-      mesh.current.geometry.boundingBox?.getCenter(center);
-      return center;
-    }
-  }, []);
-
-  useFrame((state, dt) => {
-    if (positionsRef.current) {
-      positionsRef.current.set(cpMarkers);
-      positionsRef.current.needsUpdate = true;
-    }
-    if (surfacePositionsRef.current) {
-      surfacePositionsRef.current.set(verts[chartType][blades]);
-      surfacePositionsRef.current.needsUpdate = true;
-    }
-  });
-  return (
-    <mesh ref={mesh} scale={[0.1, 6, 1]}>
-      <planeGeometry
-        args={[5, 5, 50, 60]}
-        onUpdate={(self) => {
-          self.computeVertexNormals();
-          console.log(self);
-        }}
-      >
-        <bufferAttribute
-          ref={surfacePositionsRef}
-          attach="attributes-position"
-          count={verts[chartType][blades].length / 3}
-          array={verts[chartType][blades]}
-          itemSize={3}
-        />
-      </planeGeometry>
-      <points ref={points}>
-        <bufferGeometry>
-          <bufferAttribute
-            ref={positionsRef}
-            attach="attributes-position"
-            count={cpMarkers.length / 3}
-            array={cpMarkers}
-            itemSize={3}
-            // usage={THREE.DynamicDrawUsage}
-          />
-        </bufferGeometry>
-        <pointsMaterial
-          attach="material"
-          color={"blue"}
-          size={5}
-          sizeAttenuation={false}
-        />
-      </points>
-
-      <meshStandardMaterial
-        color="lightgreen"
-        side={THREE.DoubleSide}
-        opacity={0.6}
-        transparent
-        wireframe
-      />
-      <Html
-        className="select-none"
-        color="black"
-        scale={[10, 0.1, 1]}
-        up={[0, -10, 0]}
-        position={[35, 0, 0]}
-        center
-      >
-        Angle
-      </Html>
-      <Html
-        className="select-none"
-        color="black"
-        scale={[10, 0.1, 1]}
-        position={[60, 0.2, 0]}
-        center
-      >
-        Cp
-      </Html>
-      <Html
-        className="select-none"
-        color="black"
-        scale={[10, 0.1, 1]}
-        position={[8, 0, 2.5]}
-        center
-      >
-        J
-      </Html>
-    </mesh>
-  );
-};
-
-const Lights = () => {
-  const pointLightRef = useRef(null!);
-  useHelper(pointLightRef, THREE.PointLightHelper, 100);
-  return (
-    <>
-      <pointLight ref={pointLightRef} position={[10, 10, 10]} />
-    </>
-  );
-};
+import PowerUnitResultsCp from "./PowerUnitResultsCp";
 
 interface Point {
   v: number;
@@ -152,7 +31,6 @@ const PowerUnitResults = () => {
   const speed = usePropellerStore((state) => state.cruiseSpeed);
   const blades = usePropellerStore((state) => state.blades);
   const setAltitude = useResultsStore((state) => state.setAltitude);
-  const cpMarkers = useResultsStore((state) => state.cpMarkers);
   const setCpMarkers = useResultsStore((state) => state.setCpMarkers);
 
   const [table, setTable] = useState<Point[]>();
@@ -203,20 +81,7 @@ const PowerUnitResults = () => {
           </div>
         </div>
       </div>
-      <div className="h-96 w-96">
-        <Canvas>
-          <axesHelper />
-          <ambientLight intensity={0.4} />
-          <Lights />
-          <Surface cpMarkers={cpMarkers} />
-          <OrbitControls
-            autoRotate
-            autoRotateSpeed={0.25}
-            target={[3.5, 1, 2.5]}
-          />
-          {/* <Stats /> */}
-        </Canvas>
-      </div>
+      <PowerUnitResultsCp />
     </div>
   );
 };
